Conflicts:
	components/com_fabrik/views/list/tmpl/default.xml

+++ components/com_fabrik/models/element.php
@@ -2173,13 +2173,15 @@ class PlgFabrik_Element extends FabrikPlugin
 		$element->element_ro = $this->getROElement($model->data, $c);
 		$element->value = $this->getValue($model->data, $c);
 
-		if (array_key_exists($elHTMLName . '_raw', $model->data))
+		$elName = $this->getFullName(true, false);
+		
+		if (array_key_exists($elName . '_raw', $model->data))
 		{
-			$element->element_raw = $model->data[$elHTMLName . '_raw'];
+			$element->element_raw = $model->data[$elName . '_raw'];
 		}
 		else
 		{
-			$element->element_raw = array_key_exists($elHTMLName, $model->data) ? $model->data[$elHTMLName] : $element->value;
+			$element->element_raw = array_key_exists($elName, $model->data) ? $model->data[$elName] : $element->value;
 		}
 
 		if ($this->dataConsideredEmpty($element->element_ro, $c))
@@ -2407,17 +2409,29 @@ class PlgFabrik_Element extends FabrikPlugin
 		}
 
 		$params = $this->getParams();
-		$customLink = $params->get('custom_link');
+		$customLink = $params->get('custom_link', '');
 
 		if ($customLink !== '' && $this->getElement()->link_to_detail == '1' && $params->get('custom_link_indetails', true))
 		{
 			$w = new FabrikWorker;
 
-			foreach ($data as $k => $val)
+			/**
+			 * $$$ hugh - this should really happen elsewhere, but I needed a quick fix for handling
+			 * {slug} in detail view links, which for some reason are not 'stringURLSafe' at this point,
+			 * so they are like "4:A Page Title" instead of 4-a-page-title.
+			 */
+			if (strstr($customLink, '{slug}'))
 			{
-				$repData[$k] = $val;
+				$slug = str_replace(':', '-', $data['slug']);
+				$slug = JApplication::stringURLSafe($slug);
+				$customLink = str_replace('{slug}', $slug, $customLink);
 			}
-
+			
+			/**
+			 * Testing new parseMessageForRepeats(), see comments on the function itself.
+			 */
+			$customLink = $w->parseMessageForRepeats($customLink, $data, $this, $repeatCounter);
+			
 			$customLink = $w->parseMessageForPlaceHolder($customLink, $data);
 			$customLink = $this->getListModel()->parseMessageForRowHolder($customLink, $data);
 
@@ -2524,6 +2538,8 @@ class PlgFabrik_Element extends FabrikPlugin
 
 	/**
 	 * Helper method to build an input field
+	 *
+	 * @deprecated use JLayouts instead
 	 *
 	 * @param   string  $node      Input type default 'input'
 	 * @param   array   $bits      Input property => value
@@ -3517,12 +3533,15 @@ class PlgFabrik_Element extends FabrikPlugin
 	/**
 	 * Get sub option values
 	 *
+	 * @param   array  $data  Form data. If submitting a form, we want to use that form's data and not
+	 *                        re-query the form Model for its data as with multiple plugins of the same type
+	 *                        this was getting the plugin params out of sync.
+	 *
 	 * @return  array
 	 */
-
-	protected function getSubOptionValues()
+	protected function getSubOptionValues($data = array())
 	{
-		$phpOpts = $this->getPhpOptions();
+		$phpOpts = $this->getPhpOptions($data);
 
 		if (!$phpOpts)
 		{
@@ -3560,12 +3579,15 @@ class PlgFabrik_Element extends FabrikPlugin
 	/**
 	 * Get sub option labels
 	 *
+	 * @param   array  $data  Form data. If submitting a form, we want to use that form's data and not
+	 *                        re-query the form Model for its data as with multiple plugins of the same type
+	 *                        this was getting the plugin params out of sync.
+	 *
 	 * @return  array
 	 */
-
-	protected function getSubOptionLabels()
+	protected function getSubOptionLabels($data = array())
 	{
-		$phpOpts = $this->getPhpOptions();
+		$phpOpts = $this->getPhpOptions($data);
 
 		if (!$phpOpts)
 		{
@@ -3628,12 +3650,15 @@ class PlgFabrik_Element extends FabrikPlugin
 	/**
 	 * Should we get the elements sub options via the use of eval'd parameter setting
 	 *
+	 * @param   array  $data  Form data. If submitting a form, we want to use that form's data and not
+	 *                        re-query the form Model for its data as with multiple plugins of the same type
+	 *                        this was getting the plugin params out of sync.
 	 * @since  3.0.7
 	 *
 	 * @return mixed  false if no, otherwise needs to return array of JHTML::options
 	 */
 
-	protected function getPhpOptions()
+	protected function getPhpOptions($data = array())
 	{
 		$params = $this->getParams();
 		$pop = $params->get('dropdown_populate', '');
@@ -3641,7 +3666,8 @@ class PlgFabrik_Element extends FabrikPlugin
 		if ($pop !== '')
 		{
 			$w = new FabrikWorker;
-			$pop = $w->parseMessageForPlaceHolder($pop, $this->getFormModel()->getData());
+			$data = empty($data) ? $this->getFormModel()->getData() : $data;
+			$pop = $w->parseMessageForPlaceHolder($pop, $data);
 
 			if (FabrikHelperHTML::isDebug())
 			{
@@ -5900,7 +5926,7 @@ class PlgFabrik_Element extends FabrikPlugin
 		{
 			if (!array_key_exists(0, $data))
 			{
-				// Occurs if we have created a list from an exisitng table whose data contains json objects (e.g. #__users.params)
+				// Occurs if we have created a list from an existing table whose data contains json objects (e.g. #__users.params)
 				$obj = JArrayHelper::toObject($data);
 				$data = array();
 				$data[0] = $obj;
@@ -5921,7 +5947,20 @@ class PlgFabrik_Element extends FabrikPlugin
 			$r = empty($data) ? '' : array_shift($data);
 		}
 
-		return $r;
+		$layout = $this->getLayout('list');
+		$displayData = new stdClass;
+		$displayData->text = $r;
+		$res = $layout->render($displayData);
+
+		// If no custom list layout found revert to the default list renderer
+		if ($res === '')
+		{
+			$basePath = COM_FABRIK_FRONTEND . '/layouts/';
+			$layout = new JLayoutFile('fabrik-element-list', $basePath, array('debug' => false, 'component' => 'com_fabrik', 'client' => 'site'));
+			$res = $layout->render($displayData);
+		}
+
+		return $res;
 	}
 
 	/**
@@ -5986,40 +6025,17 @@ class PlgFabrik_Element extends FabrikPlugin
 			return;
 		}
 
-		$id = $this->getHTMLId($repeatCounter);
-		$valueid = $id . '_ddVal';
-		$labelid = $id . '_ddLabel';
-		$value = '<input class="inputbox text" id="' . $valueid . '" name="addPicklistValue" />';
-		$label = '<input class="inputbox text" id="' . $labelid . '" name="addPicklistLabel" />';
-		$str[] = '<a href="#" title="' . FText::_('COM_FABRIK_ADD') . '" class="btn btn-info toggle-addoption">';
-		$str[] = FabrikHelperHTML::image('plus.png', 'form', @$this->tmpl, array('alt' => FText::_('COM_FABRIK_ADD')));
-		$str[] = '</a>';
-		$str[] = '<div style="clear:left">';
-		$str[] = '<div class="addoption"><div>' . FText::_('COM_FABRIK_ADD_A_NEW_OPTION_TO_THOSE_ABOVE') . '</div>';
+		$basePath = COM_FABRIK_BASE . '/components/com_fabrik/layouts/element';
+		$layout = new JLayoutFile('fabrik-element-addoptions', $basePath, array('debug' => false, 'component' => 'com_fabrik', 'client' => 'site'));
+		$displayData = new stdClass;
+		$displayData->id = $this->getHTMLId($repeatCounter);
+		$displayData->add_image = FabrikHelperHTML::image('plus.png', 'form', @$this->tmpl, array('alt' => FText::_('COM_FABRIK_ADD')));
+		$displayData->allowadd_onlylabel = $params->get('allowadd-onlylabel');
+		$displayData->savenewadditions = $params->get('savenewadditions');
+		$displayData->onlylabel = $onlylabel;
+		$displayData->hidden_field = $this->getHiddenField($displayData->id . '_additions', '', $displayData->id . '_additions');
 
-		if (!$params->get('allowadd-onlylabel') && $params->get('savenewadditions'))
-		{
-			// $$$ rob don't wrap in <dl> as the html is munged when rendered inside form tab template
-			$str[] = '<label for="' . $valueid . '">' . FText::_('COM_FABRIK_VALUE') . '</label>';
-			$str[] = $value;
-
-			if (!$onlylabel)
-			{
-				$str[] = '<label for="' . $labelid . '">' . FText::_('COM_FABRIK_LABEL') . '</label>';
-				$str[] = $label;
-			}
-		}
-		else
-		{
-			$str[] = $label;
-		}
-
-		$str[] = '<input class="button btn btn-success" type="button" id="' . $id . '_dd_add_entry" value="' . FText::_('COM_FABRIK_ADD') . '" />';
-		$str[] = $this->getHiddenField($id . "_additions", '', $id . "_additions");
-		$str[] = '</div>';
-		$str[] = '</div>';
-
-		return implode("\n", $str);
+		return $layout->render($displayData);
 	}
 
 	/**
@@ -6693,6 +6709,13 @@ class PlgFabrik_Element extends FabrikPlugin
 		// Swap dec and thousand seps back to Normal People Decimal Format!
 		$decimal_sep = $params->get('field_decimal_sep', '.');
 		$thousand_sep = $params->get('field_thousand_sep', ',');
+		
+		// Workaround for params not letting us save just a space!
+		if ($thousand_sep == '#32')
+		{
+			$thousand_sep = ' ';
+		}
+		
 		$val = str_replace($thousand_sep, '', $val);
 		$val = str_replace($decimal_sep, '.', $val);
 
@@ -7192,6 +7215,18 @@ class PlgFabrik_Element extends FabrikPlugin
 	{
 		$this->defaults = null;
 	}
+	
+	/**
+	 * Clear default values, need to call this if we change an elements value in any of the formData
+	 * arrays during submisison process.
+	 *
+	 * @return  null
+	 */
+	
+	public function clearDefaults()
+	{
+		$this->defaults = null;
+	}
 
 	/**
 	 * Should the 'label' field be quoted.  Overridden by databasejoin and extended classes,
@@ -7456,23 +7491,28 @@ class PlgFabrik_Element extends FabrikPlugin
 		{
 			$formData = $this->getFormModel()->formDataWithTableName;
 			$parentId = $formData[$k];
-			$query->delete($join->table_join)->where('parent_id = ' . $db->quote($parentId));
-			$db->setQuery($query);
-			$db->execute();
-		}
-
-		foreach ($idsToKeep as $parentId => $ids)
-		{
-			$query->clear();
-			$query->delete($join->table_join)->where('parent_id = ' . $parentId);
-
-			if (!empty($ids))
+			if (!empty($parentId))
 			{
-				$query->where('id NOT IN ( ' . implode($ids, ',') . ')');
+				$query->delete($join->table_join)->where('parent_id = ' . $db->quote($parentId));
+				$db->setQuery($query);
+				$db->execute();
 			}
-
-			$db->setQuery($query);
-			$db->execute();
+		}
+		else
+		{
+			foreach ($idsToKeep as $parentId => $ids)
+			{
+				$query->clear();
+				$query->delete($join->table_join)->where('parent_id = ' . $parentId);
+	
+				if (!empty($ids))
+				{
+					$query->where('id NOT IN ( ' . implode($ids, ',') . ')');
+				}
+	
+				$db->setQuery($query);
+				$db->execute();
+			}
 		}
 	}
 
@@ -7521,20 +7561,80 @@ class PlgFabrik_Element extends FabrikPlugin
 	{
 		$fbConfig = JComponentHelper::getParams('com_fabrik');
 		$params = $this->getParams();
-		
 		$advancedClass = '';
+		$globalAdvanced = (int) $fbConfig->get('advanced_behavior', '0');
 
-		if ($fbConfig->get('advanced_behavior', '0') == '1')
+		if ($globalAdvanced !== 0)
 		{
-
 			if (!$this->getGroup()->canRepeat())
 			{
-				$advancedClass = $params->get('advanced_behavior', '0') == '1'  ? 'advancedSelect' : '';
+				$advancedClass = $params->get('advanced_behavior', '0') == '1' || $globalAdvanced === 2  ? 'advancedSelect' : '';
 			}
-			
 		}
 		
 		return $advancedClass;
-		
+	}
+
+	/**
+	 * Get the element's JLayout file
+	 * Its actually an instance of FabrikLayoutFile which inverses the ordering added include paths.
+	 * In FabrikLayoutFile the addedPath takes precedence over the default paths, which makes more sense!
+	 *
+	 * @param   string  $type  form/details/list
+	 *
+	 * @return FabrikLayoutFile
+	 */
+	public function getLayout($type)
+	{
+		$name = get_class($this);
+		$name = strtolower(JString::str_ireplace('PlgFabrik_Element', '', $name));
+		$basePath = COM_FABRIK_BASE . '/plugins/fabrik_element/' . $name . '/layouts';
+		$layout = new FabrikLayoutFile('fabrik-element-' . $name. '-' . $type, $basePath, array('debug' => false, 'component' => 'com_fabrik', 'client' => 'site'));
+		$layout->addIncludePaths(JPATH_THEMES . '/' . JFactory::getApplication()->getTemplate() . '/html/layouts');
+
+		return $layout;
+	}
+
+	/**
+	 * Validate teh element against a Joomla form Rule
+	 *
+	 * @param   string  $type   Rule type e.g. 'password'
+	 * @param   mixed   $value  Value to validate
+	 * @param   mixed   $path   Optional path to load teh rule from
+	 *
+	 * @throws Exception
+	 *
+	 * @return bool
+	 */
+	protected function validateJRule($type, $value, $path = null)
+	{
+		if (!is_null($path))
+		{
+			JFormHelper::addRulePath($path);
+		}
+
+		$app = JFactory::getApplication();
+		$rule = JFormHelper::loadRuleType($type, true);
+		$xml  = new SimpleXMLElement('<xml></xml>');
+		$lang = JFactory::getLanguage();
+		$lang->load('com_users');
+
+		if (!$rule->test($xml, $value))
+		{
+			$this->validationError = '';
+
+			foreach ($app->getMessageQueue() as $i => $msg)
+			{
+				if ($msg['type'] === 'warning')
+				{
+					$this->validationError .= $msg['message'] . '<br />';
+				}
+			}
+			FabrikWorker::killMessage($app, 'warning');
+
+			return false;
+		}
+
+		return true;
 	}
 }
